added too read me and updated setup file, addtionally moved the untils around
