+++ src/devices/LoRa/rx.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 #include "mySX1278.h"
-#include "utils.h"
+#include "../utils.h"
 #include <stdio.h>
 #include <unistd.h>
