+++ src/devices/utils.c
