+++ src/devices/utils.h
